fix null exception

+++ Link-Backend-EF/Resources/PatientResource.cs
@@ -4,8 +4,8 @@
     {
         public int Id { get; set; }
         public bool Active { get; set; }
-        public Decimal Weight { get; set; }
-        public Decimal Height { get; set; }
+        public float Weight { get; set; }
+        public float Height { get; set; }
         public string Country { get; set; }
         public int UserDataId { get; set; }
     }
