+++ Link-Backend-EF/Resources/SavePatientResource.cs
@@ -7,9 +7,9 @@ namespace Link_Backend_EF.Resources
         [Required]
         public bool Active { get; set; }
         [Required]
-        public Decimal Weight { get; set; }
+        public float Weight { get; set; }
         [Required]
-        public Decimal Height { get; set; }
+        public float Height { get; set; }
         [Required]
         public string Country { get; set; }
         [Required]
